Main/requirements.txt
tensorflow
numpy
opencv-python
mediapipe
pillow